+++ src/Extensions/GFChainedSelects/Type/WPObject/FormField/FieldValue/ValueProperty.php
@@ -11,6 +11,7 @@ declare( strict_types = 1 );
 namespace WPGraphQL\GF\Extensions\GFChainedSelects\Type\WPObject\FormField\FieldValue;
 
 use WPGraphQL\AppContext;
+use WPGraphQL\GF\Model\FormField;
 use WPGraphQL\GF\Type\WPObject\FormField\FieldValue\FieldValues;
 
 /**
@@ -28,7 +29,7 @@ class ValueProperty extends FieldValues {
 				'type'        => [ 'list_of' => 'String' ],
 				'description' => __( 'ChainedSelect field value.', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ): ?array {
-					if ( ! self::is_field_and_entry( $source, $context ) ) {
+					if ( ! $source instanceof FormField || ! isset( $context->gfEntry ) ) {
 						return null;
 					}