+++ src/Type/WPObject/Entry/SubmittedEntry.php
@@ -89,7 +89,7 @@ class SubmittedEntry extends AbstractObject implements TypeWithInterfaces, Field
 			'post'           => [
 				'type'        => 'Post',
 				'description' => __( 'For forms with Post fields, this is the post object that was created.', 'wp-graphql-gravity-forms' ),
-				'resolve'     => static fn ( $source, array $args, AppContext $context ) => ! empty( $source->postDatabaseId ) ? $context->get_loader( 'post' )->load( $source->postDatabaseId ) : null,
+				'resolve'     => static fn ( $source, array $args, AppContext $context ) => ! empty( $source->postDatabaseId ) ? $context->get_loader( 'post' )->load_deferred( $source->postDatabaseId ) : null,
 			],
 			'postDatabaseId' => [
 				'type'        => 'Int',