+++ src/Type/WPObject/Order/OrderItem.php
@@ -11,11 +11,10 @@ declare( strict_types = 1 );
 namespace WPGraphQL\GF\Type\WPObject\Order;
 
 use WPGraphQL\AppContext;
-use WPGraphQL\GF\Model\FormField as FormFieldModel;
+use WPGraphQL\GF\Data\Loader\FormFieldsLoader;
 use WPGraphQL\GF\Type\Enum\CurrencyEnum;
 use WPGraphQL\GF\Type\WPInterface\FormField;
 use WPGraphQL\GF\Type\WPObject\AbstractObject;
-use WPGraphQL\GF\Utils\GFUtils;
 
 /**
  * Class - OrderItem
@@ -110,9 +109,13 @@ class OrderItem extends AbstractObject {
 				'type'        => FormField::$type,
 				'description' => __( 'The form field that the order item is connected to', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ) {
-					$field = GFUtils::get_field_by_id( $context->gfForm->form, $source['id'] );
+					if ( ! isset( $context->gfForm ) || ! isset( $source['id'] ) ) {
+						return null;
+					}
 
-					return new FormFieldModel( $field, $context->gfForm->form );
+					$id_for_loader = FormFieldsLoader::prepare_loader_id( $context->gfForm->databaseId, (int) $source['id'] );
+
+					return $context->get_loader( FormFieldsLoader::$name )->load_deferred( $id_for_loader );
 				},
 			],
 		];